Backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pandas<2.2.0
# Advanced Analytics
scikit-learn==1.4.1.post1
statsmodels==0.14.1
mlxtend==0.23.1
joblib==1.3.2

# Testing / Linting
pytest==8.0.2
pytest-flask==1.3.0
ruff==0.3.0

# Deployment
waitress>=3.0.0
python-dotenv>=1.0.0
